Fixed styles for author sales.
