+++ frontend/src/AuthorInventory.scss
@@ -1,6 +1,7 @@
 #author-page-container {
     width: 90%;
     padding-top: 100px;
+    border: solid 1px black;
 }
 
 #author-page-title {
@@ -16,6 +17,7 @@
     align-items: center;
     justify-content: space-around;
     width: 100%;
+    border: 1px solid black;
 }
 
 #author-inventory-title-container {
