+++ frontend/src/AuthorSales.scss
@@ -1,9 +1,10 @@
 #author-sales-container {
-  padding-top: 140px;
+  // padding-top: 70px;
   // box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
-  margin: 20px;
-  height: 100%;
-  width: 90%;
+  // margin: 20px;
+  margin-top: 50px;
+  height: 90%;
+  width: 98%;
 }
 
 #sales-container h2 {
@@ -20,7 +21,7 @@
 
 #author-sales-content-left {
   display: flex;
-  justify-content: space-between;
+  // justify-content: space-between;
   flex-direction: column;
   height: 100%;
   gap: 20px;
@@ -28,7 +29,6 @@
 
 #author-sales-content-right {
   flex: 1;
-
 }
 
 #total-sales, #books-sold {
@@ -44,12 +44,14 @@
 
 #total-sales h3, #books-sold h3 {
   color: #666;
-  margin-bottom: 10px;
+  // margin-bottom: 10px;
+  margin: 0;
 }
 
 #total-sales p {
   font-size: 24px;
   font-weight: bold;
+  margin: 0;
 }
 
 #books-sold ul {
