+++ frontend/src/CustomDropdown.scss
@@ -4,13 +4,13 @@
     width: 250px;
     font-family: Arial, sans-serif;
   }
-  
+
   .dropdown-button {
     display: flex;
     align-items: center;
     justify-content: space-between;
     width: 100%;
-    padding: 10px 15px;
+    padding: 7.5px 10px;
     background-color: white;
     border: 1px solid #d1d5db;
     border-radius: 4px;
@@ -21,17 +21,17 @@
     box-shadow: 0 1px 2px rgba(0, 0, 0, 0.05);
     transition: all 0.2s ease;
   }
-  
+
   .dropdown-button:hover {
     background-color: #f9fafb;
   }
-  
+
   .dropdown-button:focus {
     outline: none;
     border-color: #3b82f6;
     box-shadow: 0 0 0 3px rgba(59, 130, 246, 0.3);
   }
-  
+
   .dropdown-arrow {
     width: 20px;
     height: 20px;
@@ -39,11 +39,11 @@
     margin-left: 10px;
     transition: transform 0.2s ease;
   }
-  
+
   .dropdown-arrow.open {
     transform: rotate(180deg);
   }
-  
+
   .dropdown-menu {
     position: absolute;
     z-index: 10;
@@ -57,7 +57,7 @@
     border-radius: 4px;
     box-shadow: 0 4px 6px -1px rgba(0, 0, 0, 0.1), 0 2px 4px -1px rgba(0, 0, 0, 0.06);
   }
-  
+
   .dropdown-item {
     display: flex;
     align-items: center;
@@ -68,17 +68,17 @@
     color: #1f2937;
     transition: background-color 0.2s ease;
   }
-  
+
   .dropdown-item:hover {
     background-color: #e0f2fe;
   }
-  
+
   .dropdown-item.selected {
     background-color: #eff6ff;
     color: #2563eb;
     font-weight: 500;
   }
-  
+
   .check-icon {
     display: flex;
     align-items: center;
@@ -86,22 +86,22 @@
     width: 20px;
     height: 20px;
   }
-  
+
   /* For webkit browsers like Chrome and Safari */
   .dropdown-menu::-webkit-scrollbar {
     width: 6px;
   }
-  
+
   .dropdown-menu::-webkit-scrollbar-track {
     background: #f1f1f1;
     border-radius: 3px;
   }
-  
+
   .dropdown-menu::-webkit-scrollbar-thumb {
     background: #d1d5db;
     border-radius: 3px;
   }
-  
+
   .dropdown-menu::-webkit-scrollbar-thumb:hover {
     background: #9ca3af;
   }
